pyproject.toml
[project]
name = "lgs"
version = "0.1.0"
description = "Microservice for content moderation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.3",
    "fastapi[standard]>=0.115.14",
    "psycopg[binary,pool]>=3.2.9",
    "pyahocorasick>=2.2.0",
    "pydantic-settings>=2.10.1",
    "sentry-sdk[fastapi]>=2.32.0",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1.3",
    "ipdb>=0.13.13",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.0",
    "pytest-django>=4.11.1",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.11.3",
    "autoflake>=2.3.1",
    "debugpy>=1.8.14",
    "isort>=6.0.1",
    "pipenv>=2025.0.3",
    "pyright>=1.1.402",
    "python-lsp-server[all]>=1.12.2",
    "readline>=6.2.4.2",
    "pytest>=8.4.1",
    "cookiecutter>=2.6.0",
]

[tool.ruff.lint]
extend-select = [
    "ERA",
    "YTT",
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    "B",
    "A",
    "COM",
    "C4",
    "C4",
    "T10",
    "DJ",
    "FIX",
    "FA",
    "ISC",
    "LOG",
    "G",
    "T20",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "SLOT",
    "TID",
    "TC",
    "ARG",
    "PTH",
    "I",
    "C90",
    "N",
    "PERF",
    "E",
    "W",
    #"D", #pydocstyle
    "F",
    "PLC", #pylint convention
    "PLE", #pylint error
    "PLR", #pylint refactor
    "PLW", #pylint warning
    "UP",
    "FURB",
    "TRY",
]
[tool.ruff.lint.per-file-ignores]
"*test*.py" = ["S101", "PLR2004"]
"__init__.py" = ["F403"]
"app/migrations/*" = ["ALL"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["*_tests.py"]

[[tool.mypy.overrides]]
module = ["ahocorasick"]
ignore_missing_imports = true
